requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
llama-cpp-python>=0.2.23
pydantic>=2.5.2
python-docx>=1.0.1
PyPDF2>=3.0.1
numpy>=1.26.2
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
python-dotenv>=1.0.0
torch>=2.1.1
transformers>=4.35.2
sentence-transformers>=2.2.2
pytest>=7.4.3
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
psutil>=5.9.6 